[frontend] refactor uploader (#9760)
